+++ Changes.md
@@ -4,7 +4,9 @@
 Improvements
 ------------
 
-- Constraint : Improved performance for UV constraints where the target has static geometry but an animated transform. One such benchmark shows a greater than 40x speedup.
+- Constraint :
+  - Added `maintainReferencePosition` plug. This allows the constraint to maintain the original position of the object at a specifed reference time.
+  - Improved performance for UV constraints where the target has static geometry but an animated transform. One such benchmark shows a greater than 40x speedup.
 
 1.0.5.1 (relative to 1.0.5.0)
 =======